requirements.txt
streamlit
groq
python-dotenv
requests
pandas
hypothesis
pytest
